vex-arduino-serial-bridge/python/requirements.txt
PySerial

